Add testcase that test basic functionalities
Added testcase that check if the memories are creating correctly and if
the memories are displaying correctly.
Fixed the size of the user image on the home page.
Removed smooth centering of the map when clicked
(the map marker was displayed much later than the map shift, so it felt
like the map was freezing.)